move rearrangecoeff
