+++ lib/distributeSearch.js
@@ -8,7 +8,7 @@ const Negative = require('./util/Negative');
 const NodeCreator = require('./util/NodeCreator');
 const NodeStatus = require('./util/NodeStatus');
 const NodeType = require('./util/NodeType');
-const PolynomialTermOperations = require('./PolynomialTermOperations');
+const rearrangeCoefficient = require('./simplifyBasics/rearrangeCoefficient');
 const TreeSearch = require('./util/TreeSearch');
 
 const distributeSearch = TreeSearch.postOrder(distribute);
@@ -228,7 +228,7 @@ function simplify(node) {
   const substeps = [];
   const simplifyFunctions = [
     evaluateArithmeticSearch,                       // e.g. 2*9 -> 18
-    PolynomialTermOperations.rearrangeCoefficient,  // e.g. x*5 -> 5x
+    rearrangeCoefficient,  // e.g. x*5 -> 5x
     collectAndCombineSearch,                        // e.g 2x*4x -> 8x^2
     distributeAndSimplifyOperationNode, // e.g. (2+x)(3+x) -> 2*(3+x) recurses
   ];
