+++ lib/simplifyBasicsSearch.js
@@ -12,8 +12,9 @@ const NodeCreator = require('./util/NodeCreator');
 const NodeStatus = require('./util/NodeStatus');
 const NodeType = require('./util/NodeType');
 const PolynomialTermNode = require('./PolynomialTermNode');
-const PolynomialTermOperations = require('./PolynomialTermOperations');
 const TreeSearch = require('./util/TreeSearch');
+
+const rearrangeCoefficient = require('./simplifyBasics/rearrangeCoefficient');
 
 const SIMPLIFICATION_FUNCTIONS = [
   // multiplication by 0 yields 0
@@ -35,7 +36,7 @@ const SIMPLIFICATION_FUNCTIONS = [
   // If this is a / node and the denominator is 1 or -1, get rid of it
   removeDivisionByOne,
   // e.g. x*5 -> 5x
-  PolynomialTermOperations.rearrangeCoefficient,
+  rearrangeCoefficient,
 ];
 
 const simplifyBasicsSearch = TreeSearch.preOrder(simplifyBasics);
