+++ test/PolynomialTerm.test.js
@@ -6,7 +6,6 @@ const math = require('mathjs');
 const flatten = require('../lib/util/flattenOperands');
 const PolynomialTermNode = require('../lib/PolynomialTermNode');
 const PolynomialTermOperations = require('../lib/PolynomialTermOperations');
-const print = require('./../lib/util/print');
 
 function testIsPolynomialTerm(exprStr, isTerm) {
   it(exprStr + ' ' + isTerm, function () {
@@ -66,20 +65,3 @@ describe('canSimplifyPolynomialTerms addition', function() {
   tests.forEach(t => testCanCombine(t[0], t[1]));
 });
 
-function testRearrangeCoefficient(exprStr, outputStr) {
-  it(exprStr + ' -> ' + outputStr, function () {
-    const inputNode = flatten(math.parse(exprStr));
-    const newNode = PolynomialTermOperations.rearrangeCoefficient(inputNode).newNode;
-    assert.equal(
-      print(newNode),
-      outputStr);
-  });
-}
-
-describe('rearrangeCoefficient', function() {
-  const tests = [
-    ['2 * x^2', '2x^2'],
-    ['y^3 * 5', '5y^3'],
-  ];
-  tests.forEach(t => testRearrangeCoefficient(t[0], t[1]));
-});
